moved org_id with snet to snet folder
