+++ src/assets/thirdPartyServices/index.js
@@ -2,8 +2,8 @@ import React, { lazy } from "react";
 
 const ExampleService = lazy(() => import("./snet/example_service"));
 const CNTKImageRecognition = lazy(() => import("./snet/cntk_image_recon"));
-const CNTKNextDayTrend = lazy(() => import("./CNTKNextDayTrend.js"));
-const CNTKLSTMForecast = lazy(() => import("./CNTKLSTMForecast.js"));
+const CNTKNextDayTrend = lazy(() => import("./snet/cntk_next_day_trend"));
+const CNTKLSTMForecast = lazy(() => import("./snet/cntk_lstm_forecast"));
 const CNTKLanguageUnderstanding = lazy(() => import("./snet/cntk_language_understanding"));
 const I3DActionRecognition = lazy(() => import("./I3DActionRecognition.js"));
 const OpenNMTRomanceTranslator = lazy(() => import("./OpenNMTRomanceTranslator.js"));
@@ -21,11 +21,11 @@ const TimeSeriesAnomalyDiscoveryService = lazy(() => import("./TimeSeriesAnomaly
 const VisualQAOpencog = lazy(() => import("./VisualQAOpencog.js"));
 const MosesService = lazy(() => import("./MosesService"));
 const SemanticSegmentationService = lazy(() => import("./SemanticSegmentation.js"));
-const FaceDetectService = lazy(() => import("./FaceDetectService"));
+const FaceDetectService = lazy(() => import("./snet/face_detect"));
 const FaceLandmarksService = lazy(() => import("./FaceLandmarksService"));
-const FaceAlignService = lazy(() => import("./FaceAlignService"));
-const FaceIdentityService = lazy(() => import("./FaceIdentityService"));
-const EmotionRecognitionService = lazy(() => import("./EmotionRecognitionService"));
+const FaceAlignService = lazy(() => import("./snet/face_align"));
+const FaceIdentityService = lazy(() => import("./snet/face_identity"));
+const EmotionRecognitionService = lazy(() => import("./snet/emotion_recognition_service"));
 const HolisticEdgeDetectionService = lazy(() => import("./HolisticEdgeDetectionService"));
 const ImageRetrievalService = lazy(() => import("./ImageRetrievalService"));
 const GeneAnnotationService = lazy(() => import("./GeneAnnotationService"));
@@ -33,7 +33,7 @@ const TranslationService = lazy(() => import("./TranslationService"));
 const NewsSummaryService = lazy(() => import("./NewsSummaryService"));
 const StyleTransfer = lazy(() => import("./StyleTransfer"));
 const LanguageDetectionService = lazy(() => import("./LanguageDetectionService"));
-const CoreferenceResolutionService = lazy(() => import("./CoreferenceResolutionService"));
+const CoreferenceResolutionService = lazy(() => import("./snet/coreference_resolution_service"));
 const NamedEntityDisambiguation = lazy(() => import("./NamedEntityDisambiguation"));
 const NetworkAnalysisBipartite = lazy(() => import("./NetworkAnalysisBipartite"));
 const NetworkAnalysisRobustness = lazy(() => import("./NetworkAnalysisRobustness"));
