+++ src/assets/thirdPartyServices/snet/coreference_resolution_service/index.js
