+++ src/assets/thirdPartyServices/snet/emotion_recognition_service/index.js
@@ -1,6 +1,6 @@
 import React from "react";
-import SNETImageUpload from "./standardComponents/SNETImageUpload";
-import EmotionVisualizer from "./emotion-recognition-service/EmotionVisualizer";
+import SNETImageUpload from "../../standardComponents/SNETImageUpload";
+import EmotionVisualizer from "../../emotion-recognition-service/EmotionVisualizer";
 
 export default class EmotionRecognitionService extends React.Component {
   constructor(props) {
